refactor: infrastructure#repository로 패키지구조 추가, static 변수 위치 수정
