+++ src/main/java/roomescape/infrastructure/repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.infrastructure.repository;
 
 import java.util.List;
 import java.util.Optional;
